remove OntologyStateImpl
Another pointless subclass
